requirements.txt
pandas
numpy
streamlit
groq
sentence-transformers
faiss-cpu
scikit-learn
python-dotenv
plotly
requests